# Speedup `brew --prefix <formula>`