--- a/Library/Homebrew/test/cmd/--prefix_spec.rb
+++ b/Library/Homebrew/test/cmd/--prefix_spec.rb
@@ -8,7 +8,7 @@ describe "brew --prefix" do
 
   it "prints a given Formula's prefix", :integration_test do
     expect { brew "--prefix", testball }
-      .to output(%r{#{HOMEBREW_CELLAR}/testball}o).to_stdout
+      .to output("#{HOMEBREW_PREFIX}/opt/testball\n").to_stdout
       .and not_to_output.to_stderr
       .and be_a_success
   end