--- a/manpages/brew.1
+++ b/manpages/brew.1
@@ -997,7 +997,7 @@ Linux: \fB/home/linuxbrew/\.linuxbrew\fR
 .IP "" 0
 .
 .P
-If \fIformula\fR is provided, display the location in the Cellar where \fIformula\fR is or would be installed\.
+If \fIformula\fR is provided, display the location where \fIformula\fR is or would be installed\.
 .
 .TP
 \fB\-\-unbrewed\fR