This case is _really_ slow even although it's something we encourage people to run often and build systems often do. The `brew --prefix` case is really fast because it's just in Bash so: let's pull the `brew --prefix <formula>` case into Bash too.

This doesn't handle any edge-cases like `--installed` and the formula detection is pretty simple.

Also, to make this behaviour consistent, never output `Cellar` paths from the (Ruby) `brew --prefix`; we never want people relying on the Cellar paths anyway, only output them if the formula wasn't installed (where, arguably, using a Cellar path is even worse) and the speedup is worth this deviation in behaviour.

Benchmarks:

```console
$ hyperfine --min-runs=3 --warmup=1 "git checkout origin/master; brew --prefix" "git checkout speedup_brew_prefix; brew --prefix"                    
Benchmark #1: git checkout origin/master; brew --prefix
  Time (mean ± σ):     163.6 ms ±  21.2 ms    [User: 19.5 ms, System: 74.1 ms]
  Range (min … max):   133.8 ms … 184.6 ms    19 runs
 
Benchmark #2: git checkout speedup_brew_prefix; brew --prefix
  Time (mean ± σ):     168.0 ms ±  20.1 ms    [User: 20.3 ms, System: 77.4 ms]
  Range (min … max):   134.3 ms … 186.4 ms    18 runs
 
Summary
  'git checkout origin/master; brew --prefix' ran
    1.03 ± 0.18 times faster than 'git checkout speedup_brew_prefix; brew --prefix'

$ hyperfine --min-runs=3 --warmup=1 "git checkout origin/master; brew --prefix mysql@5.7" "git checkout speedup_brew_prefix; brew --prefix mysql@5.7"
Benchmark #1: git checkout origin/master; brew --prefix mysql@5.7
  Time (mean ± σ):      1.565 s ±  0.102 s    [User: 351.9 ms, System: 577.7 ms]
  Range (min … max):    1.482 s …  1.678 s    3 runs
 
Benchmark #2: git checkout speedup_brew_prefix; brew --prefix mysql@5.7
  Time (mean ± σ):     293.0 ms ±  19.5 ms    [User: 25.9 ms, System: 207.0 ms]
  Range (min … max):   275.2 ms … 333.3 ms    9 runs
 
Summary
  'git checkout speedup_brew_prefix; brew --prefix mysql@5.7' ran
    5.34 ± 0.50 times faster than 'git checkout origin/master; brew --prefix mysql@5.7'
```